simmodels-package(cropsim)	R Documentation

What the package does (short line) Dynamic eco(physio)logical simulation models

Description

More about what it does (maybe more than one line) Simmodels data manipulation. Particularly in the context of ecological modeling, such as crop growth simulation modeling.

Details

Package: 	 Simmodels

Type: 	 Package

Version: 	 0.1

Date: 	 2008-10-17

License: 	 GPL v3

LazyLoad: 	 yes

~~ An overview of how to use the package, including the most important functions ~~

Author(s)

Robert J. Hijmans

Maintainer: <r.hijmans@gmail.com>

References

~~ Literature or other references for background information ~~

See Also

~~ Optional links to other man pages, e.g. ~~ ~~ <pkg> ~~

[Package cropsim version 0.1 Index]